[BEAM-6165] FlinkMetricContainer cleanup (#7637)
